+++ src/app/(app)/@modals/params.tsx
@@ -1,5 +1,6 @@
 export enum ModalParam {
   AddQuest = "add-quest",
+  BuildInfo = "BuildInfo",
   QuestChains = "quest-chains",
   Stats = "stats",
 }